Add files via upload
fixed the spacing for the most part, adjusted the tooltips for both dark and light modes, adjusting the wording, reorganized the columns with the inputs to group them together more based on the type of inputs they were, split the current investment card into current and future dollars. adjusted the light mode for color impact as some colors for background or boxes were clashing with the rest of the content, not standing out. added the lines to the table view and highlighted the retirement age to show visually in the table view in yellow.